advert type service updated

+++ src/main/java/com/ph/service/AdvertTypeService.java
@@ -2,6 +2,7 @@ package com.ph.service;
 
 
 import com.ph.domain.entities.AdvertType;
+import com.ph.exception.customs.ConflictException;
 import com.ph.exception.customs.ResourceNotFoundException;
 import com.ph.payload.mapper.AdvertMapper;
 import com.ph.payload.request.AdvertTypeRequest;
@@ -134,6 +135,9 @@ public class AdvertTypeService {
     public ResponseEntity<AdvertTypeResponse> delete(Long id) {
         // Retrieve the AdvertType by its ID
         AdvertType advertType = getById(id);
+        if (!advertType.getAdverts().isEmpty()) {
+            throw new ConflictException("AdvertType cannot be deleted because it has associated Adverts");
+        }
 
         // Delete the AdvertType from the repository
         repository.delete(advertType);
